Session - 18 RoomDataBase

+++ app/src/main/java/com/example/demooflayouts/roomDBTask/TaskDetails.kt
@@ -7,8 +7,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "taskDetail")
 data class TaskDetails(
-    @PrimaryKey(autoGenerate = true) val taskId : Long,
-    @ColumnInfo(name = "task_title") val taskTitle : String,
-    @ColumnInfo(name = "task_desc") val taskDesc : String,
-    @ColumnInfo(name="email") val userEmail : String
+    @PrimaryKey(autoGenerate = true) val taskId: Int,
+    @ColumnInfo(name = "task_title") val taskTitle: String,
+    @ColumnInfo(name = "task_desc") val taskDesc: String,
+    @ColumnInfo(name = "email") val userEmail: String
 )
